added new methods for manipulating registers
Register structs are templated with a backing type and an access type
place registers into their memory addresses via placement new as before

the other new struct is RegisterAccess (name may change)
it's templated with a backing type, a value type, a mask, and an access type
its Get and Set functions are conditional on the access type

in your register subtype, define RegisterAccess types via a using statement to spec out your register's values

Register will enable/disable its Get, Set, and SetBatch functions depending on its access type

SetBatch allows us to set the full register in a single write operation by specifying the accesses and values to set. It operates on a raw value, updating it, and finally setting it after accesses have applied to it

+++ include/Display/SpriteManager.hpp
@@ -56,7 +56,7 @@ private:
         std::int32_t BeginIndex{ 0 };
         std::int32_t Length{ 0 };
 
-        std::int32_t GetTileIndex() const { return BeginIndex / CountPerTile; }
+        std::uint16_t GetTileIndex() const { return BeginIndex / CountPerTile; }
     };
 
     // All current loaded tile data by asset
